style: 스타일 변경 모바일 사이즈 적용 pretandard 폰트 적용
