+++ src/app/components/bread-crumb/page.tsx
@@ -1,0 +1,25 @@
+import React from "react";
+import BreadCrumb from "./BreadCrumb";
+
+export default function BreadCrumbPage() {
+  // 예시 데이터
+  const breadcrumbItems = [
+    { label: "홈", href: "/" },
+    { label: "컴포넌트", href: "/components" },
+    { label: "네비게이션", href: "/components/navigation" },
+    { label: "브레드크럼", href: "/components/navigation/breadcrumb", isCurrentPage: true },
+  ];
+
+  return (
+    <div className="p-8">
+      <h1 className="text-2xl font-bold mb-6">브레드크럼 컴포넌트</h1>
+
+      <div className="space-y-8">
+        <div>
+          <h2 className="text-xl font-semibold mb-4">자동 생성 브레드크럼</h2>
+          <BreadCrumb autoGenerate={true} items={[]} />
+        </div>
+      </div>
+    </div>
+  );
+}
