+++ src/types.d.ts
@@ -1,0 +1,11 @@
+/// <reference types="react" />
+
+import React from 'react';
+
+declare global {
+  namespace JSX {
+    interface IntrinsicElements {
+      [elemName: string]: any;
+    }
+  }
+}
